// src/interfaces/gridster-item.ts
export interface GridsterItem {
    x?: number;
    y?: number;
    rows?: number;
    cols?: number;
    initCallback?: Function;
    dragEnabled?: boolean;
    resizeEnabled?: boolean;
    maxItemRows?: number;
    minItemRows?: number;
    maxItemCols?: number;
    minItemCols?: number;
    minItemArea?: number;
    maxItemArea?: number;
    [propName: string]: any;
}